requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
boto3==1.18.27
gunicorn==20.1.0
python-dotenv==0.19.0
